+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/State/QuerySpec.hs
@@ -10,17 +10,10 @@
 module Test.Cardano.Ledger.Api.State.QuerySpec (spec) where
 
 import Cardano.Ledger.Api.Era
-import Cardano.Ledger.Api.State.Query (
-  CommitteeMemberState (..),
-  CommitteeMembersState (..),
-  HotCredAuthStatus (..),
-  MemberStatus (..),
-  NextEpochChange (..),
-  QueryPoolStateResult,
-  getNextEpochCommitteeMembers,
-  queryCommitteeMembersState,
- )
+import Cardano.Ledger.Api.State.Query
 import Cardano.Ledger.BaseTypes
+import Cardano.Ledger.Coin
+import Cardano.Ledger.Compactible
 import Cardano.Ledger.Conway.Governance (
   Committee (..),
   ConwayEraGov (..),
@@ -45,28 +38,39 @@ import qualified Data.Map.Strict as Map
 import Data.Maybe (isNothing)
 import Data.Set (Set)
 import qualified Data.Set as Set
+import qualified Data.VMap as VMap
 import Lens.Micro ((&), (.~), (^.))
 import Test.Cardano.Ledger.Api.Arbitrary ()
+import Test.Cardano.Ledger.Binary.Random
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Conway.Arbitrary ()
+import Test.Cardano.Ledger.Core.Arbitrary
 import Test.Cardano.Ledger.Core.Binary.RoundTrip (roundTripEraExpectation)
 import Test.Cardano.Ledger.Shelley.Arbitrary ()
 import Test.Cardano.Slotting.Numeric ()
 
 spec :: Spec
 spec = do
-  describe "API Types" $ do
-    describe "Roundtrip" $ do
-      prop "Shelley" $ roundTripEraExpectation @ShelleyEra @QueryPoolStateResult
-      prop "Allegra" $ roundTripEraExpectation @AllegraEra @QueryPoolStateResult
-      prop "Mary" $ roundTripEraExpectation @MaryEra @QueryPoolStateResult
-      prop "Alonzo" $ roundTripEraExpectation @AlonzoEra @QueryPoolStateResult
-      prop "Babbage" $ roundTripEraExpectation @BabbageEra @QueryPoolStateResult
-      prop "Conway" $ roundTripEraExpectation @ConwayEra @QueryPoolStateResult
-      prop "Dijkstra" $ roundTripEraExpectation @DijkstraEra @QueryPoolStateResult
-  describe "GetCommitteeMembersState" $ do
-    committeeMembersStateSpec @ConwayEra
-    committeeMembersStateSpec @DijkstraEra
+  latestErasSpec @ConwayEra
+  latestErasSpec @DijkstraEra
+
+latestErasSpec ::
+  forall era.
+  ( ConwayEraGov era
+  , Default (StashedAVVMAddresses era)
+  , GovState era ~ ConwayGovState era
+  , ConwayEraCertState era
+  ) =>
+  Spec
+latestErasSpec = do
+  describe "QuerySpec" $ do
+    describe (eraName @era) $
+      describe "Roundtrip" $ do
+        prop "QueryPoolStateResult" $ roundTripEraExpectation @era @QueryPoolStateResult
+        prop "StakeSnapshots" $ roundTripEraExpectation @era @StakeSnapshots
+    describe "Queries" $ do
+      committeeMembersStateSpec @era
+      queryStakeSnapshotsSpec @era
 
 committeeMembersStateSpec ::
   forall era.
@@ -77,7 +81,7 @@ committeeMembersStateSpec ::
   ) =>
   Spec
 committeeMembersStateSpec =
-  prop "CommitteeMembersState Query" $ \statusFilter -> do
+  prop "GetCommitteeMembersState" $ \statusFilter -> do
     forAll genCommittee $ \committee ->
       -- half of the committee members in the next epoch will overlap with the current ones
       forAll (genNextCommittee @era committee) $ \nextCommittee ->
@@ -449,3 +453,47 @@ queryCommitteeMembersStateNoFilters =
     Set.empty
     Set.empty
     Set.empty
+
+queryStakeSnapshotsSpec ::
+  forall era.
+  ( EraCertState era
+  , EraGov era
+  , EraStake era
+  , Default (StashedAVVMAddresses era)
+  ) =>
+  Spec
+queryStakeSnapshotsSpec =
+  describe "GetStakeSnapshots" $ do
+    prop "ssStakeSnapshots has all poolIds" $ \ss -> do
+      let
+        nes = (def :: NewEpochState era) & nesEsL . esSnapshotsL .~ ss
+        result = queryStakeSnapshots nes Nothing
+        getPoolIds =
+          Map.filter ((> 0) . spssNumDelegators) . VMap.toMap . ssStakePoolsSnapShot
+        allPoolIdsWithDelegations =
+          foldMap (Map.keysSet . getPoolIds) [ssStakeMark ss, ssStakeSet ss, ssStakeGo ss]
+        allPoolIds =
+          foldMap
+            (Map.keysSet . VMap.toMap . ssStakePoolsSnapShot)
+            [ssStakeMark ss, ssStakeSet ss, ssStakeGo ss]
+        nonZeroTotal s =
+          nonZeroOr (VMap.foldMap fromCompact (unStake (ssStake s))) (knownNonZeroCoin @1)
+        nonZeroSubTotal ssWhich =
+          nonZeroOr (foldMap ssWhich (ssStakeSnapshots result)) (knownNonZeroCoin @1)
+      subPoolIds <- uniformSubSet Nothing allPoolIds QC
+      -- Tricky bit about the query is when all pool ids are requested then ones that do not have
+      -- delegations are filtered out, while when poolIds are specified, then they are retained even
+      -- if they don't have any delegations
+      let
+        subResult = queryStakeSnapshots nes (Just subPoolIds)
+      pure @Gen $
+        conjoin
+          [ Map.keysSet (ssStakeSnapshots result) === allPoolIdsWithDelegations
+          , nonZeroSubTotal ssMarkPool === ssMarkTotal result
+          , nonZeroSubTotal ssSetPool === ssSetTotal result
+          , nonZeroSubTotal ssGoPool === ssGoTotal result
+          , ssMarkTotal result === nonZeroTotal (ssStakeMark ss)
+          , ssSetTotal result === nonZeroTotal (ssStakeSet ss)
+          , ssGoTotal result === nonZeroTotal (ssStakeGo ss)
+          , Map.keysSet (ssStakeSnapshots subResult) === subPoolIds
+          ]