+++ libs/vector-map/src/Data/VMap.hs
@@ -36,6 +36,7 @@ module Data.VMap (
   toList,
   toAscList,
   keys,
+  keysSet,
   elems,
   fromAscList,
   fromAscListN,
@@ -57,6 +58,7 @@ import Control.DeepSeq
 import Data.Aeson (FromJSON (..), FromJSONKey, ToJSON (..), ToJSONKey)
 import qualified Data.Map.Strict as Map
 import Data.Maybe as Maybe hiding (mapMaybe)
+import qualified Data.Set as Set
 import Data.VMap.KVVector (KVVector (..))
 import qualified Data.VMap.KVVector as KV
 import qualified Data.Vector as V
@@ -303,6 +305,10 @@ fold = VG.foldMap' id . valsVector . unVMap
 keys :: VG.Vector kv k => VMap kv vv k v -> [k]
 keys = VG.toList . keysVector . unVMap
 {-# INLINE keys #-}
+
+keysSet :: VG.Vector kv k => VMap kv vv k v -> Set.Set k
+keysSet = Set.fromDistinctAscList . VG.toList . keysVector . unVMap
+{-# INLINE keysSet #-}
 
 elems :: VG.Vector vv v => VMap kv vv k v -> [v]
 elems = VG.toList . valsVector . unVMap