+++ libs/cardano-ledger-api/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.13.0.0
 
+* Add `queryStakeSnapshots` and the types produced by it:
+  - `StakeSnapshot`
+  - `StakeSnapshots`
 * Re-export `DirectDeposits` and `directDepositsTxBodyL` from `Cardano.Ledger.Api.Tx.Body`.
 * Re-export `constitutionGuardrailsScriptHashL` from `Cardano.Ledger.Api.Governance`.
 * Changed the type of the following functions by adding `Network` argument: